% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes_birds_doc.R
\docType{data}
\name{bes_birds}
\alias{bes_birds}
\title{Baltimore Ecosystem: Bird Communities}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 908 rows and 28 columns.
}
\source{
{Cary Institute Of Ecosystem Studies and C. Nilon. 2017. Biodiversity, Fauna, Vacant Lot Bird Communities ver 150. Environmental Data Initiative. https://doi.org/10.6073/pasta/69ba063217dbf21a9862e95a1ea3a527 (Accessed 2021-03-17).}

\url {https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-bes.3170.150}
}
\usage{
bes_birds
}
\description{
Studying the effect of different predictors of the bird communities and different species of birds
}
\keyword{datasets}
