% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luq_snails_doc.R
\docType{data}
\name{luq_snails}
\alias{luq_snails}
\title{Luquillo Forest: Snails}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9120 rows and 26 columns.
}
\source{
{Willig, M. 2020. El Verde long-term invertebrate data - Luquillo Forest Dynamics Plot (LFDP) ver 9996737. Environmental Data Initiative. https://doi.org/10.6073/pasta/45e3a90ed462f66acdde83636746f87f (Accessed 2021-03-17).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-luq.107.9996737}
}
\usage{
luq_snails
}
\description{
Dataset containing abundance of multiple species of snails, as well as overall abundance
Hurricanes in 1989 (Hugo), 1998 (Georges), 2017 (Maria)
}
\keyword{datasets}
