% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr_coral_doc.R
\docType{data}
\name{mcr_coral}
\alias{mcr_coral}
\title{Moorea Coral Reef: LTER
Coral Reef Resilience: Live and Dead Pocillopora and Acropora Coral Colony Time Series}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 942 rows and 14 columns.
}
\source{
{MCR LTER: Coral Reef Resilience: Live and Dead Pocillopora and Acropora Coral Colony Time Series from 2006 to 2011}

{MCR LTER: Coral Reef: Long-term Population and Community Dynamics: Corals, ongoing since 2005}

\url{Moorea Coral Reef LTER, T. Adam, A. Brooks, R. Schmitt, and S. Holbrook. 2019. MCR LTER: Coral Reef Resilience: Live and Dead Pocillopora and Acropora Coral Colony Time Series from 2006 to 2011 ver 2. Environmental Data Initiative. https://doi.org/10.6073/pasta/23157cfdedcc6522c999e593d83007c9 (Accessed 2021-03-01).}
 \url{Moorea Coral Reef LTER and P. Edmunds. 2015. MCR LTER: Coral Reef: Long-term Population and Community Dynamics: Corals, ongoing since 2005 ver 33. Environmental Data Initiative. https://doi.org/10.6073/pasta/1f05f1f52a2759dc096da9c24e88b1e8 (Accessed 2021-03-12).}
}
\usage{
mcr_coral
}
\description{
Data describing abundance, size structure, morphologies of living & dead corals in 2006, 2009, 2010, 2011
Individual coral colonies identifies to genus, scored as living or dead, total area of footprint calculated
Branch morphology on a scale from 1 -> 3 with tight -> open spacing
}
\keyword{datasets}
