% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr_fish_doc.R
\docType{data}
\name{mcr_fish}
\alias{mcr_fish}
\title{Moorea Coral Reef: Changes in the abundance of fish functional groups}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 11 columns.
}
\source{
{Moorea Coral Reef LTER, T. Adam, A. Brooks, and P. Edmunds. 2016. MCR LTER: Coral Reef: Changes in the abundance of fish functional groups: Adam et al. 2014 Oecologia ver 2. Environmental Data Initiative. https://doi.org/10.6073/pasta/2cea607bcfc63e41c5eee1da15e5a112 (Accessed 2021-03-01).}

\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-mcr.1041.2}
}
\usage{
mcr_fish
}
\description{
Data generated to analyze relationship between live/dead coral and changes in abundance of functional groups of fish following coral loss.
How will coral reef fish communities respond to climate-driven disturbances?
}
\keyword{datasets}
